# Add utilities to profile netket optimisation loop