--- a/netket/driver/abstract_variational_driver.py
+++ b/netket/driver/abstract_variational_driver.py
@@ -25,7 +25,7 @@ import jax
 
 from netket.logging import AbstractLog, JsonLog
 from netket.operator._abstract_observable import AbstractObservable
-from netket.utils import mpi
+from netket.utils import mpi, timing
 from netket.utils.types import Optimizer, PyTree
 from netket.vqs import VariationalState
 
@@ -104,6 +104,7 @@ class AbstractVariationalDriver(abc.ABC):
         self._loss_stats = None
         self._loss_name = minimized_quantity_name
         self._step_count = 0
+        self._timer = None
 
         self._variational_state = variational_state
         self.optimizer = optimizer
@@ -245,6 +246,7 @@ class AbstractVariationalDriver(abc.ABC):
         callback: Callable[
             [int, dict, "AbstractVariationalDriver"], bool
         ] = lambda *x: True,
+        timeit: bool = False,
     ):
         """
         Runs this variational driver, updating the weights of the network stored in
@@ -287,6 +289,7 @@ class AbstractVariationalDriver(abc.ABC):
                 serialized to disk (ignored if logger is provided)
             write_every: Every how many steps the json data should be flushed to disk (ignored if
                 logger is provided)
+            timeit: If True, provide timing information.
         """
 
         if not isinstance(n_iter, numbers.Number):
@@ -313,51 +316,61 @@ class AbstractVariationalDriver(abc.ABC):
         callbacks = _to_iterable(callback)
         callback_stop = False
 
-        with tqdm(
-            total=n_iter,
-            disable=not show_progress,
-            dynamic_ncols=True,
-        ) as pbar:
-            old_step = self.step_count
-            first_step = True
-
-            for step in self.iter(n_iter, step_size):
-                log_data = self.estimate(obs)
-                self._log_additional_data(log_data, step)
-
-                # if the cost-function is defined then report it in the progress bar
-                if self._loss_stats is not None:
-                    pbar.set_postfix_str(self._loss_name + "=" + str(self._loss_stats))
-                    log_data[self._loss_name] = self._loss_stats
-
-                # Execute callbacks before loggers because they can append to log_data
-                for callback in callbacks:
-                    if not callback(step, log_data, self):
-                        callback_stop = True
-
-                for logger in loggers:
-                    logger(self.step_count, log_data, self.state)
-
-                if len(callbacks) > 0:
-                    if mpi.mpi_any(callback_stop):
-                        break
-
-                # Reset the timing of tqdm after the first step, to ignore compilation time
-                if first_step:
-                    first_step = False
-                    pbar.unpause()
-
-                # Update the progress bar
-                pbar.update(self.step_count - old_step)
+        with timing.timed_scope(force=timeit) as timer:
+            with tqdm(
+                total=n_iter,
+                disable=not show_progress,
+                dynamic_ncols=True,
+            ) as pbar:
                 old_step = self.step_count
+                first_step = True
 
-            # Final update so that it shows up filled.
-            pbar.update(self.step_count - old_step)
+                for step in self.iter(n_iter, step_size):
+                    with timing.timed_scope(name="observables"):
+                        log_data = self.estimate(obs)
+                        self._log_additional_data(log_data, step)
+
+                    # if the cost-function is defined then report it in the progress bar
+                    if self._loss_stats is not None:
+                        pbar.set_postfix_str(
+                            self._loss_name + "=" + str(self._loss_stats)
+                        )
+                        log_data[self._loss_name] = self._loss_stats
+
+                    # Execute callbacks before loggers because they can append to log_data
+                    for callback in callbacks:
+                        if not callback(step, log_data, self):
+                            callback_stop = True
+
+                    with timing.timed_scope(name="loggers"):
+                        for logger in loggers:
+                            logger(self.step_count, log_data, self.state)
+
+                    if len(callbacks) > 0:
+                        if mpi.mpi_any(callback_stop):
+                            break
+
+                    # Reset the timing of tqdm after the first step, to ignore compilation time
+                    if first_step:
+                        first_step = False
+                        pbar.unpause()
+
+                    # Update the progress bar
+                    pbar.update(self.step_count - old_step)
+                    old_step = self.step_count
+
+                # Final update so that it shows up filled.
+                pbar.update(self.step_count - old_step)
 
         # flush at the end of the evolution so that final values are saved to
         # file
         for logger in loggers:
             logger.flush(self.state)
+
+        if timeit:
+            self._timer = timer
+            if self._is_root:
+                print(timer)
 
         return loggers