--- a/netket/optimizer/qgt/qgt_jacobian_dense.py
+++ b/netket/optimizer/qgt/qgt_jacobian_dense.py
@@ -20,7 +20,7 @@ from jax import numpy as jnp
 from flax import struct
 
 from netket.utils.types import Scalar, PyTree
-from netket.utils import mpi
+from netket.utils import mpi, timing
 import netket.jax as nkjax
 from netket.nn import split_array_mpi
 
@@ -34,6 +34,7 @@ from .qgt_jacobian_common import (
 )
 
 
+@timing.timed
 def QGTJacobianDense(
     vstate=None,
     *,