This PR adds some coarse profiling utilities to netket.

By adding a `timeit=True` to `driver.run`, as 
```python
gs.run(n_iter=500, out="test", timeit=True)
```

We profile the total runtime as well as some inner functions in the driver. The functions that are shown in the profile are those decorated with a `@nk.utils.timing.timed`. This is very low-overhead, so for reasonable sizes and as long as we don't decorate every tiny function this will not slow down optimisation..

The idea is that this allows to see quickly, for example, if we are spending a lot of time sampling discarded elements in the Markov chain (for example, in this example we spend 20% of the sampling time sampling discarded entries), as well as seeing if we are dominated by SR or other parts of the algorithm.


```python
~/Dropbox/Ricerca/Codes/Python/netket pv/timer ⇡ 2m 13s
python-3.11.2 ❯ ipython -i Examples/Ising1d/ising1d_sr.py
Python 3.11.2 (main, Apr  7 2023, 16:35:55) [Clang 14.0.3 (clang-1403.0.22.14.1)]
Type 'copyright', 'credits' or 'license' for more information
IPython 8.15.0 -- An enhanced Interactive Python. Type '?' for help.
100%|██████████████████████████████████████████████████████████████████████████████| 500/500 [00:13<00:00, 37.75it/s, Energy=-25.4869 ± 0.0029 [σ²=0.0132, R̂=1.0044, τ=0.9<1.7]]]
╭───────────────────────────────────────────────────────────────────────────── Timing Information ──────────────────────────────────────────────────────────────────────────────╮
│ Total: 14.848                                                                                                                                                                 │
│ ├── (43.4%) | MCState.expect_and_grad : 6.440 s                                                                                                                               │
│ │   └── (67.1%) | MCState.sample : 4.318 s                                                                                                                                    │
│ │       └── (19.9%) | sampling n_discarded samples : 0.861 s                                                                                                                  │
│ ├── (53.7%) | AbstractLinearPreconditioner.__call__ : 7.976 s                                                                                                                 │
│ │   ├── (2.8%) | QGTOnTheFly : 0.222 s                                                                                                                                        │
│ │   └── (96.7%) | LinearOperator.solve : 7.714 s                                                                                                                              │
│ └── (1.1%) | loggers : 0.165 s                                                                                                                                                │
╰───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────╯
```