@PhilipVinc 2024-04-01
It will not report any timing inside of jax compiled functions. While possible, this is beyond the scope of the PR.

Moreover, timing inside of jax might degrade performance, instead this profiling here could be safely left on all the time and have a negligible impact.