@PhilipVinc 2024-04-01
It does time jitted functions, but you must time the jitted function and not jit the timed function...

https://github.com/netket/netket/blob/832a42c3dd32c48de907992426a4123c9d25b40d/netket/experimental/driver/vmc_srt.py#L38